requirements.txt
openai>=1.0.0
pytube>=15.0.0
torch>=2.0.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
streamlit>=1.22.0
opencv-python>=4.8.0
moviepy>=1.0.3
tqdm>=4.65.0 